requirements.txt
flask==3.1.0
pdfplumber==0.11.4
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
werkzeug>=3.1.5
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=12.1.1
fpdf2>=2.7.0
gunicorn==21.2.0
python-dotenv==1.0.1
openai==1.55.0
pydantic==2.10.2
azure-ai-documentintelligence==1.0.0b4
azure-core==1.32.0
flask-login

gunicorn
requests==2.31.0
